requirements.txt
asgiref==3.8.1
boto3==1.35.39
botocore==1.35.39
crispy-bootstrap5==2024.10
dj-database-url==2.2.0
Django==5.1.1
django-ckeditor-5==0.2.13
django-crispy-forms==2.3
django-environ==0.11.2
django-storages==1.14.4
django-taggit==6.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.7.0
